chore: move classes from client to correct server-side packages (Player, Board, Field) (#24)
