+++ src/main/kotlin/org/example/mankomaniaserverkotlin/model/BoardCell.kt
@@ -1,11 +1,12 @@
 /**
  * @file BoardCell.kt
- * @author 
+ * @author eles17
  * @since 
  * @description Repräsentiert ein einzelnes Spielfeld (Zelle) inkl. Position und zugewiesener Action.
  */
 package org.example.mankomaniaserverkotlin.model
-
-class BoardCell {
-    // TODO: Implementiere Eigenschaften wie Position und die zugewiesene Action
-}
+/**
+ * @property index The position of this field on the board (0-based).
+ * @property hasBranch Indicates whether the field allows the player to choose a different path (branch).
+ */
+data class BoardCell(val index:Int, val hasBranch: Boolean, val branchOptions: List<Int> = emptyList())
